[compiler] add hlo-simplify pass
