+++ compiler/include/byteir/Dialect/mhlo/Passes.h
@@ -31,6 +31,7 @@
 #include "byteir/Dialect/mhlo/Transforms/HloFolder.h"
 #include "byteir/Dialect/mhlo/Transforms/HloFuser.h"
 #include "byteir/Dialect/mhlo/Transforms/HloMove.h"
+#include "byteir/Dialect/mhlo/Transforms/HloSimplify.h"
 #include "byteir/Dialect/mhlo/Transforms/HloTransposeDotToDotGeneral.h"
 #include "byteir/Dialect/mhlo/Transforms/InsertShapeConstraint.h"
 #include "byteir/Dialect/mhlo/Transforms/LayoutTransformation.h"
