+++ compiler/include/byteir/Dialect/mhlo/Passes.td
@@ -68,7 +68,7 @@ def ConvForwardFusion : Pass<"fuse-conv-forward", "mlir::func::FuncOp"> {
 }
 
 //===----------------------------------------------------------------------===//
-// Convert Rng To CustomCall
+// Convert Op To CustomCall
 //===----------------------------------------------------------------------===//
 
 def ConvertOpToCustomCall : Pass<"convert-op-to-customcall", "ModuleOp"> {
@@ -172,6 +172,16 @@ def FusionOutlining : Pass<"fusion-outlining", "ModuleOp"> {
 def HloFolder : Pass<"hlo-fold", "mlir::func::FuncOp"> {
   let summary = "Fold mhlo ops";
   let constructor = "mlir::createHloFolderPass()";
+}
+
+//===----------------------------------------------------------------------===//
+// HloSimplify
+//===----------------------------------------------------------------------===//
+
+// TODO(liuyuanqiang): support disable patterns
+def HloSimplify : Pass<"hlo-simplify", "func::FuncOp"> {
+  let summary = "Graph simplify for hlo ops";
+  let constructor = "mlir::createHloSimplifyPass()";
 }
 
 //===----------------------------------------------------------------------===//
@@ -285,15 +295,6 @@ def LayoutTransformation : Pass<"transform-layout", "mlir::func::FuncOp"> {
 def ReduceFusion : Pass<"fuse-reduce", "mlir::func::FuncOp"> {
   let summary = "Fuse reduce and reduce-window ops";
   let constructor = "mlir::createReduceFusionPass()";
-}
-
-//===----------------------------------------------------------------------===//
-// ReshapeGather
-//===----------------------------------------------------------------------===//
-
-def ReshapeGather : Pass<"reshape-gather", "mlir::func::FuncOp"> {
-  let summary = "Reshape gather index to be 1D";
-  let constructor = "mlir::createReshapeGatherPass()";
 }
 
 //===----------------------------------------------------------------------===//
