+++ compiler/lib/Dialect/mhlo/CMakeLists.txt
@@ -98,12 +98,12 @@ add_mlir_dialect_library(ByteIRMhloPasses
   Transforms/HloFolder.cpp
   Transforms/HloMoveDown.cpp
   Transforms/HloMoveUp.cpp
+  Transforms/HloSimplify.cpp
   Transforms/HloTransposeDotToDotGeneral.cpp
   Transforms/InsertShapeConstraint.cpp
   Transforms/IOConvertFusion.cpp
   Transforms/LayoutTransformation.cpp
   Transforms/ReduceFusion.cpp
-  Transforms/ReshapeGather.cpp
   Transforms/RewriteWithConstraint.cpp
   Transforms/ShapeReification.cpp
   Transforms/StaticShapeInference.cpp
